requirements.txt
fastapi
uvicorn
tensorflow
numpy
keras
Pillow
mtcnn
opencv-python
